% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sieve_nuisances.R
\name{compute_plugin_and_IPW_sieve_nuisances}
\alias{compute_plugin_and_IPW_sieve_nuisances}
\title{Computes the data-adaptive sieve-based nuisance estimators of E[Y|A,W] and P(A=1|W)
that are constructed to be used with the IPW and plugin empirical risk functions.}
\usage{
compute_plugin_and_IPW_sieve_nuisances(
  V,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  basis_generator,
  design_function_sieve_plugin,
  weight_function_sieve_plugin,
  design_function_sieve_IPW,
  weight_function_sieve_IPW,
  family_for_targeting = binomial(),
  debug = FALSE
)
}
\arguments{
\item{V}{A matrix of observations of a subset of the covariates `W` for which to estimate the (possibly semi-marginalized) log relative risk (LRR).
The sieve basis will only be generated for these covariates.}

\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}

\item{basis_generator}{A basis_generator object (see \code{fourier_basis} and \code{bspline_basis}) ...
that specifies a sieve basis transformation of the design matrix \code{W}.}

\item{debug}{...}

\item{family}{An R \code{family} object that specifies the link and risk function used in the data-adaptive sieve update step.
`family` should be a \code{binomial} object if `Y` is binary and a \code{poisson} object if `Y` is non-binary and nonnegative (e.g. a count).}
}
\description{
Computes the data-adaptive sieve-based nuisance estimators of E[Y|A,W] and P(A=1|W)
that are constructed to be used with the IPW and plugin empirical risk functions.
}
