% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimizers_and_risk_functions.R
\name{efficient_risk_function}
\alias{efficient_risk_function}
\title{The double-robust one-step efficient empirical risk function for the log relative risk.
Note this risk function is non-convex.}
\usage{
efficient_risk_function(
  theta,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  efficient_loss_function,
  debug = FALSE,
  return_loss = FALSE,
  V = NULL,
  oracle = FALSE
)
}
\arguments{
\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}

\item{debug}{...}

\item{return_loss}{Boolean for whether to return loss function values or the risk value (i.e. average of the losses)}

\item{ERM}{A vector or matrix of log relative risk (LRR) estimates whose risk is to be evaluated using the one-step efficient double-robust risk function.}

\item{W}{A matrix of covariate observations}
}
\description{
The double-robust one-step efficient empirical risk function for the log relative risk.
Note this risk function is non-convex.
}
