% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_likelihood.R
\name{estimate_initial_likelihood}
\alias{estimate_initial_likelihood}
\title{Function to compute initial estimates of nuisance functions.}
\usage{
estimate_initial_likelihood(
  W,
  A,
  Y,
  weights = NULL,
  sl3_Learner_pA1W,
  sl3_Learner_EYAW,
  folds = 10,
  outcome_type = NULL
)
}
\arguments{
\item{W}{A column-named matrix of baseline variables.}

\item{A}{A binary vector with values in {0,1} encoding the treatment assignment.}

\item{Y}{A numeric vector storing the outcome values.}

\item{sl3_Learner_pA1W}{A \code{sl3_Learner} object from the \code{tlverse/sl3} R github package that specifies the machine-learning algorithm for learning the propensity score `P(A = 1 | W)`}

\item{sl3_Learner_EYAW}{A \code{sl3_Learner} object from the \code{tlverse/sl3} R github package that specifies the machine-learning algorithm for learning the outcome conditional mean `E[Y | A, W]`. NOTE: the treatment arms are pooled in the regression. See the preprocessing sl3_Learner \code{Lrnr_stratified} if you wish to stratify the estimation by treatment.}

\item{folds}{A number representing the number of folds to use in cross-fitting or a fold object from the package \code{tlverse/origami}. This parameter will be passed to internal \code{sl3_Task} objects that are fed to the code{sl3_Learner}s.}
}
\description{
Function to compute initial estimates of nuisance functions.
}
