% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npRRWorkingModel.R
\name{npRRWorkingModel}
\alias{npRRWorkingModel}
\title{Function to compute initial estimates of nuisance functions.}
\usage{
npRRWorkingModel(
  formula_LRR,
  W,
  A,
  Y,
  weights,
  EY1W,
  EY0W,
  pA1W,
  sl3_Learner_EYAW,
  sl3_Learner_pA1W,
  folds = 10,
  outcome_type = NULL
)
}
\arguments{
\item{W}{A column-named matrix of baseline variables.}

\item{A}{A binary vector with values in {0,1} encoding the treatment assignment.}

\item{Y}{A numeric vector storing the outcome values.}

\item{weights}{A numeric vector of weights}

\item{EY1W}{An optional numeric vector of initial estimates of `E[Y | A = 1, W]` for each individual.}

\item{EY0W}{An optional numeric vector of initial estimates of `E[Y | A = 0, W]` for each individual.}

\item{sl3_Learner_EYAW}{A \code{sl3_Learner} object from the \code{tlverse/sl3} R github package that specifies the machine-learning algorithm for learning the outcome conditional mean `E[Y | A, W]`. NOTE: the treatment arms are pooled in the regression. See the preprocessing sl3_Learner \code{Lrnr_stratified} if you wish to stratify the estimation by treatment.}

\item{sl3_Learner_pA1W}{A \code{sl3_Learner} object from the \code{tlverse/sl3} R github package that specifies the machine-learning algorithm for learning the propensity score `P(A = 1 | W)`}

\item{folds}{A number representing the number of folds to use in cross-fitting or a fold object from the package \code{tlverse/origami}. This parameter will be passed to internal \code{sl3_Task} objects that are fed to the code{sl3_Learner}s.}

\item{outcome_type}{Internal use only.}

\item{Formula_LRR}{A formula specifying the working model for the log relative risk (or msm for the log relative risk)}
}
\description{
Function to compute initial estimates of nuisance functions.
}
