% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_LRR_using_sieve.R
\name{subset_best_sieve}
\alias{subset_best_sieve}
\title{For each LRR learner, the best sieve is chosen by minimizing the double-robust one-step efficient empirical risk function over the choices of sieve-spaces.}
\usage{
subset_best_sieve(
  trained_learner_list,
  learner_names,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  efficient_loss_function,
  V = NULL
)
}
\arguments{
\item{trained_learner_list}{A unnested list of trained LRR learners as}

\item{learner_names}{Names of LRR learners. Should be `learner_names <- sapply(list_of_learners, `[[`, "name)`.}

\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}
}
\description{
For each LRR learner, the best sieve is chosen by minimizing the double-robust one-step efficient empirical risk function over the choices of sieve-spaces.
}
