% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_LRR_using_sieve_for_cv_folds.R
\name{subset_best_sieve_all_folds}
\alias{subset_best_sieve_all_folds}
\title{For each LRR learner and fold, the best sieve is chosen by minimizing the double-robust one-step efficient empirical risk function over the choices of sieve-spaces.}
\usage{
subset_best_sieve_all_folds(
  folds,
  trained_learner_list,
  learner_names,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  DR_loss_function
)
}
\arguments{
\item{folds}{A \code{folds} object (list of \code{fold} objects) from the \link{origami} package specifying the fold-specific training set at which to estimate the LRR.}

\item{trained_learner_list}{A unnested list of trained LRR learners as}

\item{learner_names}{Names of LRR learners. Should be `learner_names <- sapply(list_of_learners, `[[`, "name)`.}

\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.
This should include all observations including both the training and validation observations.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.
This should include all observations including both the training and validation observations.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}
}
\description{
For each LRR learner and fold, the best sieve is chosen by minimizing the double-robust one-step efficient empirical risk function over the choices of sieve-spaces.
}
