% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_LRR_using_sieve.R
\name{train_learners}
\alias{train_learners}
\title{Train the LRR learners for all sieves and base learners on the full data.}
\usage{
train_learners(
  V,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  family_risk_function,
  outcome_function_plugin,
  weight_function_plugin,
  outcome_function_IPW,
  weight_function_IPW,
  transform_function,
  design_function_sieve_plugin,
  weight_function_sieve_plugin,
  design_function_sieve_IPW,
  weight_function_sieve_IPW,
  family_for_targeting,
  list_of_learners,
  list_of_sieves,
  Vpred = V,
  compute = TRUE
)
}
\arguments{
\item{V}{A matrix of covariate observations.}

\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}

\item{list_of_learners}{A list of untrained \code{sl3_Learner} objects to be used to estimate the log relative risk LRR using the function \link{estimate_LRR_using_ERM}.}

\item{list_of_sieves}{A list of basis_generator objects specifying the sieve. See, for example, \code{fourier_basis} for an example and template.}

\item{Vpred}{A matrix of covariates observations at which to predict the LRR. By default, \code{Vpred} equals \code{W}.}

\item{compute}{Whether to `compute` the list of delayed trained learners. See the package \code{delayed} for details.}

\item{list_of_sieve_nuisances}{A list of sieve nuisance estimates as returned by the function \link{compute_plugin_and_IPW_sieve_nuisances}.}
}
\description{
Train the LRR learners for all sieves and base learners on the full data.
}
