% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_LRR_using_sieve_for_cv_folds.R
\name{train_learners_using_fold}
\alias{train_learners_using_fold}
\title{Train the LRR learners for all sieves and base learners on the training set specified by a fold object.}
\usage{
train_learners_using_fold(
  fold,
  V,
  A,
  Y,
  EY1W,
  EY0W,
  pA1W,
  weights,
  family_risk_function,
  outcome_function_plugin,
  weight_function_plugin,
  outcome_function_IPW,
  weight_function_IPW,
  transform_function,
  design_function_sieve_plugin,
  weight_function_sieve_plugin,
  design_function_sieve_IPW,
  weight_function_sieve_IPW,
  family_for_targeting,
  list_of_learners,
  list_of_sieves,
  Vpred = V
)
}
\arguments{
\item{fold}{\code{fold} object from the \link{origami} package specifying the fold-specific training set at which to estimate the LRR.}

\item{V}{A matrix of observations of a subset of the covariates `W` for which to estimate the (possibly semi-marginalized) log relative risk (LRR).
This should include all observations including both the training and validation observations.}

\item{A}{A binary vector specifying the treatment assignment. The values should be in {0,1}.
This should include all observations including both the training and validation observations.}

\item{Y}{A numeric vector of binary or nonnegative observations of the outcome variable.
This should include all observations including both the training and validation observations.}

\item{EY1W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=1,W] for all observations.}

\item{EY0W}{A numeric vector containing initial cross-fitted estimates of E[Y|A=0,W] for all observations.}

\item{pA1W}{A numeric vector containing initial cross-fitted estimates of P(A=1|W) for all observations.}

\item{weights}{A numeric vector of observation weights. If no special weighting desired, supply a vector of 1's.}

\item{list_of_learners}{A list of untrained \code{sl3_Learner} objects to be used to estimate the log relative risk LRR using the function \link{estimate_LRR_using_ERM}.}

\item{list_of_sieves}{A list of basis_generator objects specifying the sieve. See, for example, \code{fourier_basis} for an example and template.}
}
\description{
Train the LRR learners for all sieves and base learners on the training set specified by a fold object.
}
